+++ lib/expression.dart
@@ -1,5 +1,9 @@
+import 'package:parser_builder/combinator.dart';
+
 import 'builder_helper.dart' as helper;
 import 'parser_builder.dart';
 
 part 'src/expression/binary_expression.dart';
 part 'src/expression/identifier_expression.dart';
+part 'src/expression/postfix_expression.dart';
+part 'src/expression/prefix_expression.dart';